The Grand Auto Selector
Added the PathAutoSelector, capable of taking a starting position, a
scoring area priority, and a cross field modifier. It will then select
a path based on the ARCADE state and execute it.

Overhauled the Pathing Subsystem to allow for enum path selection.
Implemented a proper code structure.

Created a new set of SendableChoosers, which now allow for drop down
section of enums.

Old files where copied and labeled “OLD_”, will too be deprecated.